requirements.txt
dj-rest-auth==2.1.3
Django==3.1.6
django-allauth==0.44.0
django-cors-headers==3.7.0
djangorestframework==3.12.2